restructure to avoid build errors
