+++ packages/spezi-firebase-fhir/src/resources/device.ts
@@ -8,6 +8,7 @@
 
 import { type Device } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   annotationSchema,
   backboneElementSchema,
@@ -15,7 +16,6 @@ import {
   contactPointSchema,
   domainResourceSchema,
   elementSchema,
-  FhirDomainResource,
   identifierSchema,
   quantitySchema,
   referenceSchema,
