+++ packages/spezi-firebase-fhir/src/resources/diagnosticReport.ts
@@ -8,13 +8,13 @@
 
 import { type DiagnosticReport } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   attachmentSchema,
   backboneElementSchema,
   codeableConceptSchema,
   domainResourceSchema,
   elementSchema,
-  FhirDomainResource,
   identifierSchema,
   periodSchema,
   referenceSchema,
