+++ packages/spezi-firebase-fhir/src/resources/encounter.ts
@@ -8,13 +8,13 @@
 
 import { type Encounter } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   backboneElementSchema,
   codeableConceptSchema,
   codingSchema,
   domainResourceSchema,
   elementSchema,
-  FhirDomainResource,
   identifierSchema,
   intSchema,
   periodSchema,
