+++ packages/spezi-firebase-fhir/src/resources/medicationDispense.ts
@@ -8,6 +8,7 @@
 
 import { type MedicationDispense } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   annotationSchema,
   backboneElementSchema,
@@ -17,7 +18,6 @@ import {
   domainResourceSchema,
   dosageSchema,
   elementSchema,
-  FhirDomainResource,
   identifierSchema,
   quantitySchema,
   referenceSchema,
