+++ packages/spezi-firebase-fhir/src/resources/observation.ts
@@ -8,13 +8,13 @@
 
 import { type Observation } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   booleanSchema,
   codeableConceptSchema,
   dateTimeSchema,
   domainResourceSchema,
   elementSchema,
-  FhirDomainResource,
   identifierSchema,
   instantSchema,
   intSchema,
