+++ packages/spezi-firebase-fhir/src/resources/patient.ts
@@ -8,6 +8,7 @@
 
 import { type Patient } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   addressSchema,
   attachmentSchema,
@@ -19,7 +20,6 @@ import {
   dateTimeSchema,
   domainResourceSchema,
   elementSchema,
-  FhirDomainResource,
   humanNameSchema,
   identifierSchema,
   intSchema,
