+++ packages/spezi-firebase-fhir/src/resources/questionnaire.ts
@@ -8,6 +8,7 @@
 
 import { type Questionnaire, type QuestionnaireItem } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   attachmentSchema,
   backboneElementSchema,
@@ -20,7 +21,6 @@ import {
   decimalSchema,
   domainResourceSchema,
   elementSchema,
-  FhirDomainResource,
   identifierSchema,
   intSchema,
   markdownSchema,
