+++ packages/spezi-firebase-fhir/src/resources/questionnaireResponse.ts
@@ -11,6 +11,7 @@ import {
   type QuestionnaireResponseItem,
 } from 'fhir/r4b.js'
 import { z, type ZodType } from 'zod'
+import { FhirDomainResource } from './domainResourceClass.js'
 import {
   attachmentSchema,
   backboneElementSchema,
@@ -21,7 +22,6 @@ import {
   decimalSchema,
   domainResourceSchema,
   elementSchema,
-  FhirDomainResource,
   identifierSchema,
   intSchema,
   quantitySchema,
